core: Split adapter into module and extend documentation for adapter, queue and window
